Remove the two old map components and rename the version 2 map component back to map component.
