+++ leaseslicensing/frontend/leaseslicensing/src/components/internal/approvals/dashboard.vue
@@ -72,7 +72,7 @@
 <script>
 import FormSection from '@/components/forms/section_toggle.vue';
 import ApprovalsTable from '@/components/common/table_approvals.vue';
-import MapComponent from '@/components/common/component_map_with_filters_v2';
+import MapComponent from '@/components/common/component_map.vue';
 
 export default {
     name: 'InternalApprovalsDashboard',
