+++ leaseslicensing/frontend/leaseslicensing/src/components/internal/dashboard.vue
@@ -128,7 +128,7 @@ import FormSection from '@/components/forms/section_toggle.vue';
 import ApplicationsTable from '@/components/common/table_proposals';
 import ApplicationsReferredToMeTable from '@/components/common/table_proposals';
 import CompetitiveProcessesTable from '@/components/common/table_competitive_processes';
-import MapComponent from '@/components/common/component_map_with_filters_v2';
+import MapComponent from '@/components/common/component_map.vue';
 import { api_endpoints } from '@/utils/hooks';
 
 export default {
